Rename logExpansion to logTable
