+++ src/components/logdisplay/LogTable.vue
